CLEANUP: optionally skip includes/lib etc stripping
In order to repeatedly call build-shotcut.sh for building, either through
qmake+make from shotcut, or by setting envvars to make build-shotcut build
only changed files, build files need to be available. disabling CLEANUP will
make it skip this delete step.

+++ scripts/build-shotcut.sh
@@ -20,6 +20,7 @@ SOURCE_DIR="$INSTALL_DIR/src"
 ACTION_GET_COMPILE_INSTALL=1
 ACTION_GET_ONLY=0
 ACTION_COMPILE_INSTALL=1
+CLEANUP=1
 SOURCES_CLEAN=1
 INSTALL_AS_ROOT=0
 CREATE_STARTUP_SCRIPT=1
@@ -1975,7 +1976,7 @@ End-of-desktop-file
     cmd rm -rf Shotcut/Shotcut.app/share/man
   fi
   cmd tar -cjvf "$tarball" Shotcut
-  cmd rm -rf Shotcut
+  test "1" = "$CLEANUP" && cmd rm -rf Shotcut
   popd
 }
